requirements.txt
pytest==3.7.3
requests==2.21.0
pytest-allure-adaptor
